backend/requirements.txt
flask>=2.3
flask-sqlalchemy>=3.0
pymysql>=1.1
python-dotenv>=1.0
